+++ src/store/index.ts
@@ -1,3 +1,4 @@
+export * from "./backup";
 export * from "./database";
 export * from "./errors";
 export * from "./schema";